Using int64_t specializations for `ivf_pq` and `refine` (#1325)
Since FAISS and Milvus are both using `int64_t` everywhere and we haven't explicitly encountered places where `uint64_t` is being used for ann indices, we're opting to just use `int64_t` instead.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Ben Frederickson (https://github.com/benfred)

URL: https://github.com/rapidsai/raft/pull/1325

+++ cpp/src/distance/neighbors/specializations/ivfpq_build_int8_t_int64_t.cu
@@ -24,7 +24,7 @@ namespace raft::neighbors::ivf_pq {
                                raft::device_matrix_view<const T, IdxT, row_major> dataset) \
     ->index<IdxT>;
 
-RAFT_MAKE_INSTANCE(float, uint64_t);
+RAFT_MAKE_INSTANCE(int8_t, int64_t);
 
 #undef RAFT_MAKE_INSTANCE
